LTNC25_OnTap_15 err 8/10 long long fixed

+++ LTNC25_OnTap_15.cpp
@@ -3,7 +3,7 @@
 #include <cmath>
 using namespace std;
 int x ; 
-double S(int n) {
+long long S(int n) {
 	if (n == 0) return 1 ; 
 	return pow(x , n) + S(n-1);
 }
